requirements.txt
python-gitlab
PyGithub
py-gitea
requests
